bugfix for create_other_synth
need to cast it explicitly to np.uint8 before saving or we get weirdness

+++ foveated_metamers/create_other_synth.py
@@ -233,7 +233,10 @@ def main(base_image, image_to_mix, target_err, learning_rate, max_iter=100,
                title=f'Final value = {data[-1]}')
     if save_path is not None:
         mixed_image = mix_images(base_image, image_to_mix, alphas[-1], direction)
-        mixed_image = po.to_numpy(mixed_image).squeeze()
+        # if we don't convert explicitly to np.uint8 before saving, it may do
+        # funny things like saying the min value in the image should be
+        # considered 0, even if it's much larger
+        mixed_image = po.to_numpy(mixed_image).squeeze().astype(np.uint8)
         imageio.imwrite(save_path, mixed_image)
         fig.savefig(save_path.replace('.png', '_synth.svg'))
     return alphas[-1], fig
